requirements.txt
bitsandbytes
dnspython==2.6.1
djongo
djangorestframework==3.15.1
django-cors-headers
llama-index-readers-mongodb
llama-index
pymongo==4.6.3
python-dotenv==1.0.1
transformers
langchain
accelerate
requests
torch
fairscale
fire
sentencepiece
asgiref==3.8.1
Django==5.0.3
sqlparse==0.4.4
tzdata==2024.1
beautifulsoup4~=4.12.3